+++ src/env/node/git/localGitProvider.ts
@@ -502,6 +502,10 @@ export class LocalGitProvider implements GitProvider, Disposable {
 				supported = await this.git.isAtLeastVersion('2.17.0');
 				this._supportedFeatures.set(feature, supported);
 				return supported;
+			case Features.StashOnlyStaged:
+				supported = await this.git.isAtLeastVersion('2.35.0');
+				this._supportedFeatures.set(feature, supported);
+				return supported;
 			default:
 				return true;
 		}